remove scrollable in css and update ui for dndkit
